+++ lib/src/op_result_base.dart
@@ -6,15 +6,12 @@ import 'op_result_error.dart';
 /// - `E` is an enum representing possible error types.
 class OpResult<T, E extends Enum> {
   final T? _data;
-  final OpResultError<E>? _error;
+  final List<OpResultError<E>> _errors;
   final bool _isSuccess;
 
   /// Creates a success result containing `data`.
-  OpResult.success({required T data})
-    : _data = data,
-      _error = null,
-      _isSuccess = true {
-    if (data == null && !_isNullable<T>()) {
+  OpResult.success(this._data) : _errors = [], _isSuccess = true {
+    if (_data == null && !_isNullable<T>()) {
       throw ArgumentError(
         "data cannot be null when success is true and T is non-nullable. Type: ${T.toString()}",
         "data",
@@ -22,11 +19,28 @@ class OpResult<T, E extends Enum> {
     }
   }
 
-  /// Creates a failure result containing an `OpResultError`.
-  OpResult.failure({required OpResultError<E> error})
-    : _data = null,
-      _error = error,
-      _isSuccess = false;
+  /// Factory constructor for failure, allowing either a single error or multiple errors.
+  factory OpResult.failure(dynamic errorOrErrors) {
+    List<OpResultError<E>> errors;
+
+    if (errorOrErrors is OpResultError<E>) {
+      errors = [errorOrErrors];
+    } else if (errorOrErrors is List<OpResultError<E>>) {
+      if (errorOrErrors.isEmpty) {
+        throw ArgumentError("errors cannot be empty");
+      }
+      errors = errorOrErrors;
+    } else {
+      throw ArgumentError(
+        "failure() must be called with an OpResultError<E> or List<OpResultError<E>>",
+      );
+    }
+
+    return OpResult._failure(errors);
+  }
+
+  /// Private named constructor for failure cases.
+  OpResult._failure(this._errors) : _data = null, _isSuccess = false;
 
   /// Returns `true` if the result is successful.
   bool get isSuccess => _isSuccess;
@@ -42,17 +56,18 @@ class OpResult<T, E extends Enum> {
     return _data!;
   }
 
-  /// Retrieves the error, if available.
-  OpResultError<E> get error {
-    if (!isFailure) {
-      throw StateError("Attempted to access error on a successful OpResult.");
-    }
-    return _error!;
-  }
+  /// Retrieves all errors.
+  List<OpResultError<E>> get errors => _errors;
+
+  /// Retrieves the first error (for standard single-error scenarios).
+  OpResultError<E> get error => _errors.first;
+
+  @override
+  String toString() =>
+      isSuccess
+          ? "Success($data)"
+          : "Failure(${errors.map((e) => e.toString()).join(', ')})";
 
   /// Utility to check if a type is nullable.
   static bool _isNullable<T>() => null is T;
-
-  @override
-  String toString() => isSuccess ? "Success($data)" : "Failure($error)";
 }